feature: Add a student to a class
The feature is currently implemented but does not work fine. When we add a student and save the changes, the db provider throws an exception saying 'FOREIGN KEY CONSTRAINT failed'.
I noticed that when I try to execute the same set of methods in the Seed method, it works. But when I move the logic to the proper CommandHandler, it no longer works. Which is quite strange.